Retry IIB build if a specific error is encountered
This functionality was originally implemented for iiblib [1], but it has
to be removed [2] because it doesn't work and has a side effect or raising
IIBRecoverableError which disrupts the expected workflow (causes the
error details to not be printed). I think that pubtools-iib is a better
place for these retries because iiblib was not designed for blocking
calls, and such functionality is out of scope of that library.

[1] https://github.com/release-engineering/iiblib/pull/57
[2] https://github.com/release-engineering/iiblib/pull/66